Separated classes into own files
