+++ Database/Utils.py
@@ -1,0 +1,3 @@
+import uuid
+def gen_uuid():
+    return str(uuid.uuid4())
